show seekbar value -> increase in percent show Double values

+++ app/src/main/java/de/ingoreschke/hourlysalary/SalaryIncreasePresenter.kt
@@ -2,7 +2,7 @@ package de.ingoreschke.hourlysalary
 
 class SalaryIncreasePresenter(private val salaryIncreaseInteractor: SalaryIncreaseInteractor) {
 
-    fun calculateSalaryIncrease(salary: Double, increasePercentage: Int): Double {
-        return salaryIncreaseInteractor.calcIncreasedSalary(salary, increasePercentage.toDouble())
+    fun calculateSalaryIncrease(salary: Double, increasePercentage: Double): Double {
+        return salaryIncreaseInteractor.calcIncreasedSalary(salary, increasePercentage)
     }
 }
